Move InitiateDisconnectReason to its own file
